+++ valohai_cli/commands/execution/run/__init__.py
@@ -1,0 +1,3 @@
+from .frontend_command import run
+
+__all__ = ['run']
